+++ server/controllers/api/index.ts
@@ -7,6 +7,7 @@ import { usersRouter } from './users'
 import { accountsRouter } from './accounts'
 import { videosRouter } from './videos'
 import { badRequest } from '../../helpers/express-utils'
+import { videoChannelRouter } from './video-channel'
 
 const apiRouter = express.Router()
 
@@ -15,6 +16,7 @@ apiRouter.use('/oauth-clients', oauthClientsRouter)
 apiRouter.use('/config', configRouter)
 apiRouter.use('/users', usersRouter)
 apiRouter.use('/accounts', accountsRouter)
+apiRouter.use('/video-channels', videoChannelRouter)
 apiRouter.use('/videos', videosRouter)
 apiRouter.use('/jobs', jobsRouter)
 apiRouter.use('/ping', pong)